requirements.txt
agno==1.6.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
anaconda-anon-usage==0.5.0
annotated-types==0.6.0
anyio==4.9.0
archspec==0.2.3
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boltons==24.1.0
Brotli==1.0.9
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
chromadb==1.0.12
click==8.2.1
coloredlogs==15.0.1
conda==25.1.1
conda-anaconda-telemetry==0.1.2
conda-anaconda-tos==0.1.2
conda-content-trust==0.2.0
conda-libmamba-solver==25.1.1
conda-package-handling==2.4.0
conda_package_streaming==0.11.0
cryptography==43.0.3
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.16
duckdb==1.1.3
durationpy==0.10
fastapi==0.115.9
filelock==3.18.0
flatbuffers==25.2.10
frozendict==2.4.2
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.0
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.7
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.23
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
libmambapy==2.0.5
markdown-it-py==2.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.0
menuinst==2.2.0
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.42.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.86.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
orjson==3.10.18
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.3.0
phidata==2.7.10
pillow==11.2.1
pip==25.0
platformdirs==3.10.0
pluggy==1.5.0
posthog==4.10.0
propcache==0.3.2
protobuf==5.29.5
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycosat==0.6.6
pycparser==2.21
pydantic==2.10.3
pydantic_core==2.27.1
pydantic-settings==2.9.1
pydeck==0.9.1
Pygments==2.15.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.45.3
streamlit==1.41.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
truststore==0.10.0
typer==0.16.0
typing_extensions==4.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
